@@ -18,16 +18,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
+    "@angular/animations": "^14.0.2",
     "@angular/cdk": "^13.3.1",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
+    "@angular/common": "^14.0.2",
+    "@angular/compiler": "^14.0.2",
+    "@angular/core": "^14.0.2",
+    "@angular/forms": "^14.0.2",
     "@angular/material": "^13.3.1",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/platform-browser": "^14.0.2",
+    "@angular/platform-browser-dynamic": "^14.0.2",
+    "@angular/router": "^14.0.2",
     "file-saver": "^2.0.5",
     "rxjs": "7.x",
     "scss-reset": "^1.2.2",
@@ -36,14 +36,14 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "13.x",
-    "@angular-devkit/build-angular": "13.x",
+    "@angular-devkit/build-angular": "14.0.2",
     "@angular-eslint/builder": "13.x",
     "@angular-eslint/eslint-plugin": "13.x",
     "@angular-eslint/eslint-plugin-template": "13.x",
     "@angular-eslint/schematics": "13.x",
     "@angular-eslint/template-parser": "13.x",
-    "@angular/cli": "13.x",
-    "@angular/compiler-cli": "13.x",
+    "@angular/cli": "14.0.2",
+    "@angular/compiler-cli": "14.0.2",
     "@types/file-saver": "x",
     "@types/jasmine": "x",
     "@types/node": "x",
